Post first practice presentation

+++ db/migrate/20211028192609_create_ratings.rb
@@ -3,8 +3,8 @@ class CreateRatings < ActiveRecord::Migration[6.1]
     create_table :ratings do |t|
       t.integer :tutorial_id
       t.integer :user_id
-      t.integer :uxp_rating
-      t.integer :content_rating
+      t.float :uxp_rating
+      t.float :content_rating
       t.text :description
 
       t.timestamps
